/pathfinding-system:

+++ Assets/j_project/Scripts/Node.cs
@@ -16,9 +16,9 @@ namespace jfaulkner
         public int gridPosX;
         public int gridPosY;
 
-        public double G { get; set; }               // Distance from start to this node
-        public double H { get; set; }               // Heuristic (guess) distance from this node to goal
-        public double F { get; set; }               // Distance from start + heuristic distance
+        public float G { get; set; }               // Distance from start to this node
+        public float H { get; set; }               // Heuristic (guess) distance from this node to goal
+        public float F { get; set; }               // Distance from start + heuristic distance
 
 
         public List<Node> neighbors = new List<Node>();         // All succesors to node
@@ -33,15 +33,15 @@ namespace jfaulkner
 
         public void Calculate_G_Cost(Node other)
         {
-            G = other.GetG() + Math.Sqrt((Xpos - other.Xpos) * (Xpos - other.Xpos) + (Ypos - other.Ypos) * (Ypos - other.Ypos));
+            G = (float)(other.GetG() + Math.Sqrt((Xpos - other.Xpos) * (Xpos - other.Xpos) + (Ypos - other.Ypos) * (Ypos - other.Ypos)));
         }
 
-        public double GetG()
+        public float GetG()
         {
             return G;
         }
 
-        public double GetH()
+        public float GetH()
         {
             return H;
         }
